Update README.md (#5)
* Update README.md

Polybase is a web3 native database that understands tokens, wallets, and ZK. We are a drop-in replacement for Firebase, Supabase, etc with added features like public verifiability, end-end encryption, self-sovereign data & censorship resistance. 

Our mission is to restore humanity's control of its information. I would love to have Polybase as a part of your awesome list.

* Update README.md